+++ sdk/storage/azure-storage-file-share/src/main/java/com/azure/storage/file/share/ShareDirectoryAsyncClient.java
@@ -626,7 +626,8 @@ public class ShareDirectoryAsyncClient {
         return this.azureFileStorageClient.directorys().forceCloseHandlesWithRestResponseAsync(shareName, directoryPath,
             handleId, null, null, snapshot, false, context)
             .map(response -> new SimpleResponse<>(response,
-                new CloseHandlesInfo(response.getDeserializedHeaders().getNumberOfHandlesClosed())));
+                new CloseHandlesInfo(response.getDeserializedHeaders().getNumberOfHandlesClosed(),
+                    response.getDeserializedHeaders().getNumberOfHandlesFailedToClosed())));
     }
 
     /**
@@ -648,8 +649,9 @@ public class ShareDirectoryAsyncClient {
     public Mono<CloseHandlesInfo> forceCloseAllHandles(boolean recursive) {
         try {
             return withContext(context -> forceCloseAllHandlesWithTimeout(recursive, null, context)
-                .reduce(new CloseHandlesInfo(0),
-                    (accu, next) -> new CloseHandlesInfo(accu.getClosedHandles() + next.getClosedHandles())));
+                .reduce(new CloseHandlesInfo(0, 0),
+                    (accu, next) -> new CloseHandlesInfo(accu.getClosedHandles() + next.getClosedHandles(),
+                        accu.getFailedHandles() + next.getFailedHandles())));
         } catch (RuntimeException ex) {
             return monoError(logger, ex);
         }
@@ -664,7 +666,8 @@ public class ShareDirectoryAsyncClient {
                     response.getStatusCode(),
                     response.getHeaders(),
                     Collections.singletonList(
-                        new CloseHandlesInfo(response.getDeserializedHeaders().getNumberOfHandlesClosed())),
+                        new CloseHandlesInfo(response.getDeserializedHeaders().getNumberOfHandlesClosed(),
+                             response.getDeserializedHeaders().getNumberOfHandlesFailedToClosed())),
                     response.getDeserializedHeaders().getMarker(),
                     response.getDeserializedHeaders()));