+++ sdk/storage/azure-storage-file-share/src/main/java/com/azure/storage/file/share/models/CloseHandlesInfo.java
@@ -14,6 +14,7 @@ import com.azure.storage.file.share.ShareFileAsyncClient;
 public class CloseHandlesInfo {
 
     private final Integer closedHandles;
+    private final Integer failedHandles;
 
     /**
      * Creates an instance of information about close handles.
@@ -22,6 +23,18 @@ public class CloseHandlesInfo {
      */
     public CloseHandlesInfo(Integer closedHandles) {
         this.closedHandles = closedHandles;
+        this.failedHandles = 0; // What should default be??
+    }
+
+    /**
+     * Creates an instance of information about close handles.
+     *
+     * @param closedHandles The numbers of handles closed.
+     * @param failedHandles The numbers of handles that failed to close.
+     */
+    public CloseHandlesInfo(Integer closedHandles, Integer failedHandles) {
+        this.closedHandles = closedHandles;
+        this.failedHandles = failedHandles;
     }
 
     /**
@@ -30,5 +43,13 @@ public class CloseHandlesInfo {
      */
     public int getClosedHandles() {
         return this.closedHandles;
+    }
+
+    /**
+     *
+     * @return The number of handles that failed to close.
+     */
+    public int getFailedHandles() {
+        return this.failedHandles;
     }
 }