+++ sdk/storage/azure-storage-file-share/src/test/java/com/azure/storage/file/share/DirectoryAPITests.groovy
@@ -386,6 +386,7 @@ class DirectoryAPITests extends APISpec {
 
         then:
         handlesClosedInfo.getClosedHandles() == 0
+        handlesClosedInfo.getFailedHandles() == 0
         notThrown(ShareStorageException)
     }
 
@@ -410,6 +411,7 @@ class DirectoryAPITests extends APISpec {
         then:
         notThrown(ShareStorageException)
         handlesClosedInfo.getClosedHandles() == 0
+        handlesClosedInfo.getFailedHandles() == 0
     }
 
     def "Create sub directory"() {