+++ sdk/storage/azure-storage-file-share/src/test/java/com/azure/storage/file/share/FileAPITests.groovy
@@ -637,6 +637,7 @@ class FileAPITests extends APISpec {
 
         then:
         handlesClosedInfo.getClosedHandles() == 0
+        handlesClosedInfo.getFailedHandles() == 0
         notThrown(ShareStorageException)
     }
 
@@ -661,6 +662,7 @@ class FileAPITests extends APISpec {
         then:
         notThrown(ShareStorageException)
         handlesClosedInfo.getClosedHandles() == 0
+        handlesClosedInfo.getFailedHandles() == 0
     }
 
     def "Get snapshot id"() {